Use lrv- and love- prefixes; wrap <> for unknown
Fix undocumented love types (e.g., DisplayOrientation) from being marked
with lrv- prefix. Now, those types are wrapped with <> to show that they
don't have help. Love types are prefixed with love- to match their tag.

LUA_TYPES is pulled from [the manual
](http://www.lua.org/manual/5.4/manual.html#2.1) and those types are the
same in 5.1 and 5.4. Added lightuserdata which is used in love docs but
unfortunately adds a space so we need to remap to match the lrv tag.